move debug gui to new file
